Assignment_4_APIs/requirements.txt

flask
requests
mysql-connector-python
